==> Code/dallin_annotations/09_classifier_testing.Rmd <==
---
title: "Classifier Testing"
output: github_document
---
# Overview
These are my annotations to the "classifier-testing.R" script written by Adriel.

This document loads the model, and uses a predict function to test the model on new test tweets. We can see how the model will respond to tweets we find on the internet as well as tweets we make up ourselves. 

### What scripts need to function for this code to work?
##### predict_func.R
* Renamed as 11_predict_func.Rmd in this folder.
* FOUND
* WORKING

##### TweetComplaintPredictor.r
* Renamed as 12_tweet_complaint_predictor.Rmd in this folder.
* FOUND
* NOT WORKING

### Which models need to be loaded for this code to work?
##### prcomp4.dat 
* I think this is the PCR model (pr[inciple]comp[onent].dat?)
* FOUND in Code/complaint-prediction/models/fullClassifierOne/prcomp.dat, as stated in the code.
* I moved the file to the /Temporary folder

##### model50-tweet_sample_2M_noRT.dat.bin 
* It looks like this is a fasttext model, but I'm not sure if it is the supervised or unsupervised.
* FOUND
* NOT WORKING
* It must have been large enough that it couldn't be shared on github
* Found in the classifier_construction file (which is dependent on "R/clus_func.R" and "R/logit_classifier_func.R"

##### xgb.modelPC4.dat
* This must be the xg boost model that brings the sub-models together.
* FOUND in Code/complaint-prediction/models/fullClassifierOne/xgb.modelPC4.dat, as stated in the code.
* Moved to the /Temporary Folder


# Libraries
```{r}
library(fastrtext)
library(Matrix)
library(xgboost)
source("../complaint-prediction/R/predict_func.Rmd") # Formerly "R/predict_func.R" in Adriel's original code
source("tweet_complaint_predictor.Rmd") # Formerly "R/TweetComplaintPredictor.r" in Adriel's original code
```

##### Reminder of the WD
```{r}
getwd()
```


# Load Models
```{r}
load("../../Temporary/prcomp4.dat") # Formerly "complaint-prediction/models/fullClassifierOne/prcomp4.dat" in Adriel's original code
fasttextModel <- load_model("complaint-prediction/models/fullClassifierOne/model50-tweet_sample_2M_noRT.dat.bin")
# Model50 was built in fasttext_clust_EDA

load("../../Temporary/xgb.modelPC4.dat") # Formerly "complaint-prediction/models/fullClassifierOne/xgb.modelPC4.dat" in Adriel's original code
TweetComplaintPredictor <- newTweetComplaintPredictor(fasttextModel, train.pr, bst, PC4)

```

# Test Classifier
Using this predict() function, we can make predictions about single tweets. All you have to do is type your tweet in the quotations for the second argument. 
```{r}
predict(TweetComplaintPredictor, "")

```
