==> Code/dallin_annotations/12_tweet_complaint_predictor.Rmd <==
---
title: "Tweet Complaint Predictor"
output: github_document
---

# Overview
These are my annotations on the "TweetComplaintPredictor.r" script written by Adriel

```{r}
# TweetComplaintPredictor class constructor ----------------------------------------
# Class constructor 
newTweetComplaintPredictor <- function(fasttextModel, pcaModel, xgbModel, pcaModelComponentCutoff) {
  component <- enquo(pcaModelComponentCutoff)
  
  if(class(fasttextModel)[1] != "Rcpp_fastrtext"){
    stop("fasttextModel is not a member of class Rcpp_fastrtext")
  }
  if(class(pcaModel) != "prcomp"){
    stop("pcaModel is not a member of class Rcpp_fastrtext")
  }
  if(class(xgbModel) != "xgb.Booster"){
    stop("xgbModel is not a member of class Rcpp_fastrtext")
  }
  
  newTweetComplaintPredictor <- list(fasttextModel = fasttextModel,
                                     pcaModel = pcaModel,
                                     xgbModel = xgbModel,
                                     pcaModelComponentCutoff = component)
  
  class(newTweetComplaintPredictor) <- "TweetComplaintPredictor"
  
  return(newTweetComplaintPredictor)
}

# Class checker
is.TweetComplaintPredictor <- function(x) inherits(x, "TweetComplaintPredictor")



```
