==> Code/dallin_annotations/99_tweet_object.Rmd <==
---
title: "Tweet Object"
output: github_document
---
# Overview
These are my annotations on the "tweetObject.R" script written by Adriel.

```{r}
ObjectSample <- list(
  "_id" = 'ObjectId("534769eee4b0efdb7b081251")',
  "created_at" = "Fri Apr 11 04=04=57 +0000 2014",
  "id" = 'NumberLong("454470130765135872")',
  "id_str" = "454470130765135872",
  "text" = "@united is there a service fee to get the miles back if i don't clear? your rules are unclear.",
  "source" = "web",
  "truncated" = FALSE,
  "in_reply_to_status_id" = 'NumberLong("454467945172074497")',
  "in_reply_to_status_id_str" = "454467945172074497",
  "in_reply_to_user_id" = 260907612,
  "in_reply_to_user_id_str" = "260907612",
  "in_reply_to_screen_name" = "united",
  "user" = list(
    "id" = 770550,
    "id_str" = "770550",
    "name" = "Rakesh Agrawal",
    "screen_name" = "rakeshlobster",
    "location" = "San Francisco, CA",
    "url" = "http=//redesignmobile.com",
    "description" = "Global strategy at PayPal. As seen on CNBC, Bloomberg TV and VentureBeat.",
    "protected" = FALSE,
    "followers_count" = 9034,
    "friends_count" = 1672,
    "listed_count" = 410,
    "created_at" = "Tue Feb 13 21=36=08 +0000 2007",
    "favourites_count" = 4807,
    "utc_offset" = -25200,
    "time_zone" = "Pacific Time (US & Canada)",
    "geo_enabled" = TRUE,
    "verified" = FALSE,
    "statuses_count" = 39454,
    "lang" = "en",
    "contributors_enabled" = FALSE,
    "is_translator" = FALSE,
    "is_translation_enabled" = FALSE,
    "profile_background_color" = "0099B9",
    "profile_background_image_url" = "http=//abs.twimg.com/images/themes/theme4/bg.gif",
    "profile_background_image_url_https" = "https=//abs.twimg.com/images/themes/theme4/bg.gif",
    "profile_background_tile" = FALSE,
    "profile_image_url" = "http=//pbs.twimg.com/profile_images/378800000722918071/ce9e53447ed1ee7d90f5e0d1a3b800c5_normal.jpeg",
    "profile_image_url_https" = "https=//pbs.twimg.com/profile_images/378800000722918071/ce9e53447ed1ee7d90f5e0d1a3b800c5_normal.jpeg",
    "profile_link_color" = "0099B9",
    "profile_sidebar_border_color" = "5ED4DC",
    "profile_sidebar_fill_color" = "95E8EC",
    "profile_text_color" = "3C3940",
    "profile_use_background_image" = TRUE,
    "default_profile" = FALSE,
    "default_profile_image" = FALSE,
    "following" = NA,
    "follow_request_sent" = NA,
    "notifications" = NA
    ),
  "geo" = NA,
  "coordinates" = NA,
  "place" = NA,
  "contributors" = NA,
  "retweet_count" = 0,
  "favorite_count" = 0,
  "entities" = list(
    "hashtags" = list(NA),
    "symbols" = list(NA),
    "urls" = list(NA),
    "user_mentions" = list(
        "screen_name" = "united",
        "name" = "United",
        "id" = 260907612,
        "id_str" = "260907612",
        "indices" = list(0,7)
        )
    ),
  "favorited" = FALSE,
  "retweeted" = FALSE,
  "filter_level" = "medium",
  "lang" = "en",
  "sentiment" = 1,
  "date" = 'ISODate("2014-04-11T04=04=57Z")'
)


```
